+++ core-implementation/core.ts
@@ -1,6 +1,6 @@
 import grpc from "@grpc/grpc-js";
 import { addService, type GRPCFunc, loadProtoService } from "../proto.js";
-import { helloworld } from "../generated/hello.js";
+import { hello } from "./hello/hello.js";
 
 export default function addModule(server: grpc.Server) {
 	addService(
@@ -10,17 +10,3 @@ export default function addModule(server: grpc.Server) {
 		hello,
 	);
 }
-
-const hello: GRPCFunc<helloworld.HelloRequest, helloworld.HelloReply> = (
-	request,
-	respond,
-) => {
-	const message = `hello, ${request.request.name}!!!`;
-	console.log(message);
-	respond(
-		null,
-		new helloworld.HelloReply({
-			message,
-		}),
-	);
-};
