+++ package.json
@@ -19,7 +19,8 @@
 		"typescript": "^5.6.3"
 	},
 	"scripts": {
-		"build": "tsc -p . && cp -r proto dist",
+		"build": "npm run clean && tsc -p . && cp -r proto dist",
+		"clean": "rm -rf dist",
 		"dev": "tsx --watch main.ts",
 		"test": "npm run build && node --experimental-vm-modules node_modules/jest/bin/jest.js --rootDir dist"
 	},
